[eslint] Add rule against repeated keys
This rule is somewhat of a placeholder for the database asserting that there are no duplicate messages, but makes it easier for validators to fail without breaking builds.

Long term, I'd like to make this part of the database itself, but handling the error case gracefully is fairly difficult (i.e., how should duplicates be handled in watch mode? bail early? just emit errors?).