Enforce uniqueness and improve general database error handling (#23)
This is a pretty far-reaching PR that started from the aim of generating
errors when a message key is defined more than once. This ends up being
pretty complicated to handle, because there are times where a message is
_allowed_ to be redefined (i.e., a watch-mode file update that replaces
a key), but even at the same time there are cases where they are _not_
allowed to be redefined (i.e., a watch-mode file update, like any new
file, cannot define the same message more than once within itself).
There are probably ways this could be simplified, such as removing all
definitions from a source file before inserting new ones from a
recreated file, but the plumbing is already there for the current
method, so it's been kept.

However, the main issue here is that previously any kind of
`DatabaseError` during processing would end up _throwing_ that error,
leaving the operation early and leaving the database in an unknown state
of progress. This was fine, because the system would just throw entirely
and cause consumers to crash, but that's not a great experience either.
So, this PR inverts all of that and makes all processing operations
"crash free", meaning that so long as there isn't a panic internally,
all errors during processing are instead handled gracefully and returned
in an `errors` array, whether it's from parsing the source file or a
semantic error like a duplicated message trying to be inserted.
Consumers can then handle this using the `succeeded` flag on the
processing result created by `intl-loader-core`.

That opens up a lot more possibilities, including reporting on the same
kind of processing validations in ESLint through the
`native-validations` utility, which has been added as a
`no-duplicate-message-keys` rule.

Most of the work here is actually just cleaning up some tangential
behavior, like making `file_position` a required property for
`MessageValue` and plumbing through the new `SourceFileInsertionData`
rather than just returning a file key from all of the processing
methods.