[database][loader] Add `strategy` parameter for all processes
Callers can now specify the `strategy` as either creating, updating, or _replacing_ messages, with the last of those being a new operation type introduced here as well.

Unlike `update`, which requires new values to be originating from the same source file as the existing ones to allow an insertion, `replace` allows _any_ value to be written regardless of its origin. This is important for "import like" operations which bring values from an external source, like JSON from a translation vendor, and apply them to the database to be re-written back to the source files during export.

This only really affects translations, since definitions should never really need this behavior, but consistency is helpful for the future here, too.