[loader] Fix watcher not outputting precompiled files.
Previously, for some reason, there were two instances of writing precompiled definitions files here, one using the loader processing methods, and another using `db.*` directly. The latter is meh, and we want to go through the loader processing as much as possible, so I removed it. But I didn't realize that the processing version set `undefined` for the output file path, meaning it doesn't actually write to the disk.

I'm also not at all sure if that was somehow relevant to why there were two compilations happening, but there definitely only _needs_ to be a single one, so this just sets that output path properly again and gets the file written to disk.