[database][loader] Add `getDefinitionsFilesForTranslationsPath`
In projects with a lot of source files (which grows quickly with many languages supported. we're at about 350 right now), the `findAllDefinitionsFilesForTranslations` method implemented directly in `intl-loader-core` ends up being _really_ slow because it has to request the full `SourceFile` object from the native database, just to look at the `translationsPath` value. When there are also many keys in the file, this gets _really_ expensive. Instead, it's really easy to just have a dedicated method on the database to lookup translations paths on all source files and build a much shorter vector to return as the single conversion into the JS VM.

In our bundles, this brought down processing time from ~60ms _per file_ to imperceptible time (<1ms) per file.