[markdown][js] Handle template newline escapes properly
SWC's representation of template strings has a `raw` version that is the literal original text, _with escapes_, and a `cooked` version that has processed and handled escapes as needed.

We were using the `raw` version before, but it turns out that something like ``\n`` turns into `'\\n'`, which breaks newline handling downstream since it is considered as an escaped slash followed by an `n` literal, not a newline.

While I originally thought this was some issue with how the markdown parser handled escaped newlines in inline content, it turned out to just be the source, which thankfully made it much easier to resolve.