[markdown] Ignore ICU quote escape syntax (`'{`)
This syntax causes more issues than it helps, since it's common for a developer to write a message like `You entered '{name}'`, which with quote escapes would not actually interpolate the `name` variable into the message.

The real solution to this on the _user_ end is to use proper unicode quote marks, but since we can't directly control what people write, it's better to be permissive here and ensure correctness later on.