[perf] Add benchmarking test for flat-json-parser
This just sets up a benchmark for testing `intl-flat-json-parser` against serde. The difference is generally negligible since the dominating factor is just that there are a lot of bytes to process, but this can at least help identify regressions if they happen, or check simd performance in the future, if we go that route.