Rewrite Markdown parser with a DST syntax tree (#27)
This is effectively a complete rewrite of the Markdown parser to use a
true Concrete Syntax Tree implemented in the style of Rowan/Biome, where
nodes are Dynamically Sized Types (DSTs) containing the slice of their
child elements directly within the node struct. The flexibility of this
approach is immediately apparent and makes a lot of the implementation
"cleaner" in the sense that there is a single `SyntaxElement` API with
extensions, rather than a trait like `TokenSpan` and others that needs
implementing across multiple types.

This also better delineates the process of interpreting Markdown, where
the actual Syntax Tree is the concrete tree, and what _was_ called the
"AST" before is actually a compiled result of _processing_ the tree. The
result is a `CompiledElement`, which can then be passed to any of the
new `formatters`, which also brings together the previously-diasporic
implementations of icu, html, and json formatting.

Another improvement to maintining this library is the inclusion of
`xtask` and real code generation over proc macros. Using `ungram`, the
Markdown syntax is now entirely defined in `xtask/data/markdown.ungram`
and generated with `cargo xtask codegen-syntax` into standalone modules
with matching trait implementations and other behavior that was becoming
incredibly unwieldy in proc-macro land, as well as difficult to trace
when errors occurred.

Other additions/changes:
- Vastly simplified Delimiters, no longer needing multiple buffer events
and arbitrary cursor math to implement Emphasis.
- A proper `TreeBuilder` implementation that abstracts an API for
creating the tree rather than directly pushing and finishing events
within the parser itself.
- Refactored `Visit` traits to be automatically generated and enable
implementing a Fold trait in the future.
- New `TextPointer` struct acts as a flyweight handle to a piece of
text, allowing for merging text spans and efficiently cloning data with
a single API that abstracts out any necessary clones.
- Less state tracking within the Lexer, where the Parser is instead able
to inspect surrounding characters when necessary rather than inspecting
on every single token.
- Many, many more minor changes to the Parser API for consistency.
 
Unfortunately, processing speed is almost doubled in real-world tests,
even though the benchmarks have actually _improved_ performance. My
guess is that this is related to memory allocation and slight bits of
extra processing rather than the parser being necessarily slower by
design. Since the old implementation was so heavily optimized, I think
that performance can mostly be regained as this new structure gets
optimized as well.