[markdown] Fix unsafe parsing and re-add tests
The unsafe_variables tests weren't being included in the test module, so they never ran while testing the new parser implementation. As it turns out, the lexer was implemented incorrectly and would fail all of these tests, so this change both re-adds the tests and fixes the lexer to handle unsafe variables properly.