Fix missing `offset` in plural serialization.
In the refactor, I didn't include `offset` in the compiled AST structure since we don't currently process it from messages, but the AST format _expects_ it to exist in the structure for plurals (`[type, name, options, _offset_, pluralKind]`). This just fills that in with a `0` by default now to make sure the structure matches exactly what it used to look like and avoid runtime changes.