Add `text_offset` information to `SyntaxNode` and `SyntaxToken`.
This is a really slim alternative to using a proper red/green tree, where we assume that the tree is immutable and so the parser is able to insert static text offsets for every element immediately.

Ideally, this would be done in a way that adding/removing/modifying any node in the tree correctly updates the offsets of all other nodes downstream of it, but that gets very, complicated very quickly. This solution is good enough for _textual_ autofixes in diagnostics, which is the current goal.