report accurate spans in diagnostics.
Using the new `source_position`s for all syntax elements, we can accurately report sub-positions within a message for all diagnostics. This involves a few translation steps, including byte-offset to character-index to be compatible with JS/ESLint's location system, and adding some additional context to diagnostics to be able to map from a message-relative offset to a file-absolute character index in the final ESLint diagnostic.

One step closer to full autofixes.