Implement original-source offsets for escaped characters.
The motivation here is documented in the declaration of `SourceOffsetList`, but in short, writing JS or JSON messages with escaped characters like `\n` means that the _understood_ source position from inside the Markdown parser will not match the _actual_ source position, since those escaped character are processed ahead of time (otherwise the parser would have to understand how to escape strings from multiple programming languages, which is completely untenable). This misalignment would lead to incorrect reporting locations for lint errors, which downstream could cause autofixes to completely corrupt a message on accident because the reported position from the parser points to a different character in the actual source text.

The fix for this implemented here involves keeping a list of the size of every escape sequence in a source message, then using that list to compute a "true" source position when reporting back to external integrations. This was already happening in a sense for converting Rust byte-specific string indices to JS character-specific string indices, so this is really just an extension of that.

Quick testing locally shows that now all diagnostics are being reported on the correct characters in the source, though there is no technical guarantee that this is accurate yet, and there are no tests for it, either (would require some integration-level testing that doesn't currently exist).