Add more tests for `convert_byte_span_to_character_span`.
There was a bug that the end position of the span wasn't being set if it referred to the index at the end of the string, because the `<=` condition wouldn't be met when iterating characters. Now it ensures that the span is set to the end if that happens, and the added tests also ensure that both multibyte characters and escape sequences are respected properly.