fix offset adjustment calculations
The `SourceOffsetList` can be both a byte-offset _and_ a char-offset list, but realistically is more of the latter. When converting from byte to char spans, the source being referenced is the _processed_ source, meaning it has none of the offsets present in it, but also means that comparing to the offset-applied indices will give incorrect results (_sometimes_, even though it's mostly correct, surprisingly). Instead, calculating the char position first and _then_ applying the offsets ensures that all indices align with the source that's being compared before converting into the original text indices.