Fix diagnostic span alignment, offsets, and handle utf-16.
From running in the real world on our codebase, there were a number of cases of incorrect spans being reported in messages. Some were because the spans were starting from 1 or 0 incorrectly, others because the first escape offset entry wasn't being added to the list, and even more others because UTF-16 handling is not consistent between Rust and JS and there's some awkward conversion that has to be done.

After all of the changes in this commit, the fix applications are passing full verification on our codebase and have fairly high confidence that they should be accurate in all cases.