[markdown] Allow `{name}!!` as a valid ICU expression
Previously, the parser would decide whether to expect a `}` or `}!!` to close an ICU expression based on whether it was opened with `{` or `!!{`, but that invalidates a completely normal expression like `{name}!!` for no real reason. Now the parser will use the hint of the opening character to figure out how to lex the closing marker so that `}!!` won't get lexed as a single token if the expression wasn't opened with a `!!{`.