Compile documents on demand
Most things don't actually need the fully-compiled document; validators run on the CST, and the database really only needs to store the raw content and uses the parsed CST to efficiently store variables ahead of time. By removing the CST from the stored objects, we can save almost 600MB of total reserved memory (somehow, not totally sure why it's that high), and compiling on demand has the same performance impact as compiling ahead of time, since realistically we only compile a document once _maybe_ twice in a run.