[runtime] Avoid duplicate keys by passing prefixes through and tagging each node key.
Without this, some items end up just having the key `0` or `1`, and when used in an array render they can conflict with other entries in the same list. Now the absolute minimal key that will exist is `.0`, and for the case that was conflicting before it'll be `.tag-0` and `.0` for the individual items, meaning there shouldn't be any conflict.