[validator] Use real LDML plural information when enforcing plural correctness.
This adds some codegen to create a HashMap of locale to supported plural selectors that rules can then query into when determining how to enforce selectors inside of messages.

For example, we were previously enforcing that `=0` should become `zero` in English...but `zero` is not a valid English selector! It just uses `other`. So the rule was incorrect and would create possibly-failing/crashing code as a result.

Now the validator should only suggest switching to category selectors if they are valid in the locale the message exists in.