[validator] Add NoInvalidPluralSelector rule
This rule prevents usage of plural selectors that are not valid in the locale of the message being validated. For example, `ja` only has `other` as a plural selector, since there is no plural vs singular form. When writing a plural, then, only the `other` form is needed (and arguably could always be removed for simplicity, but consistency with other languages is also useful). This rule will ensure that no other categorical selectors are used in that locale (though exact selectors will always be allowed in every locale, since they aren't dependent on the LDML rules).