[validator] Add NoLimitedPlurals rule
This rule enforces the usage of `select` rather than `plural` when the list of options is not all-encompassing (i.e., only has exact selectors and no `other` catch-all clause).

Using `select` allows the type generator to provide highly-specific types that guarantee consumers are supplying only the values that are valid for the select, rather than `plural` which can only reduce the input to "any number" and leaves the possibility of unmatched values from being passed as arguments.

This also adds an `apply_fixes` method in diagnostics that can apply any set of non-overlapping diagnostic fixes, enabling validator tests to ensure the output of a fix is correct along with the reporting span and messages.