[variables] Add `NumericEnum` variant for `select` expressions with only numeric selectors
There's some ambiguity between when a `select` and a `plural` are appropriate. For consistency, we're going to enforce that any plural with only exact numeric selectors becomes a `select` instead (also allowing `other` to be carried between them). Until now, `select` has always been typed as an `Enum` with string options, but that doesn't really match the usage in this case of numeric selectors like `1`, `5`, and `10`. To address that, this adds `NumericEnum` as an automatically-detected variable type that results in `number` values being generated rather than `string` values in the final typescript definitions file.