[syntax] add `EqIgnoreSpan` trait
This trait allows for comparing any two syntax elements regardless of their location in the source code, for example meaning that two different plural arms can be compared for equality within the same source message.

This is most useful for validation rules that want to prevent duplicates or check for equality in other ways.