[eslint] Add bounds checking for reported spans
When implementing multiple projects, we encountered an issue where ESLint's cwd is the root of a monorepo, and multiple projects each independently defined the same messages a few times. Because we try to find all files in the "directory", both definitions are picked up as if they're part of the same project, leading to `AlreadyDefined` errors being thrown for each of them, incorrectly.

Additionally, something in the diagnostics for those messages causes the `file` to the original file, rather than the one being checked by eslint at the moment. This doesn't make sense given the current implementation that always references `replacement`, but it's apparently an issue for now.

To get around this, we'll do some bounds checking and log noisily when this issue occurs to ensure that diagnostics are still presented properly and warnings are shown, but without crashing the process.

Especially in subprocess workloads, these crashes can cause ESLint to run indefinitely without ever showing the reason the failure occurred.