[eslint] Fix out of bounds index access for AlreadyDefined error messages
I fFinally traced this to the error creation when I noticed that the `end` property was being set to a really high number, `content.length`. Because `messageLine` and `messageCol` are also set, this means that the computed position for the diagnostic ends up being the message position _plus_ the length of the entire file, which is obviously out of bounds. Changing that to just be length `1` means we'll still get a squiggly error in the right spot, and it should never be able to go out of the range of the file.