[eslint] Allow multiple projects to be checked simultaneously based on `projectDirectory`.
This removes the dependency on `intl-loader-core` and instead constructs a new database for every unique `projectDirectory` that's discovered while running.

When working in a monorepo, it's common to run ESLint and other linters on the entire repository together, but when using a singleton like we were with the `database` from intl-loader-core, the rules start to conflict, like `AlreadyDefined` looking at definitions in separate projects as if they conflict with each other. Using a separate database instance per configuration means there should never be cross-project conflicts.

Note that this plugin still caches validations per-file, so shared files can potentially still encounter the same issues, though that's unlikely and still a valid error if it's encountered.