Upgrade SWC and other dependencies for new Rspack version compatibility
SWC has released 40 major versions in the last 10 months..none of which really explain anything that's changed, other than some internal structures that break compatibility with other consumers of SWC like Rspack.

To maintain our own compatibility and be able to upgrade bundlers, we need to update SWC, but we also can't do that until we upgrade Rspack. It's a frustrating circular position to be in, but there shouldn't be any new work here until that upgrade is done.